Move the publication gate where both readers can reach it
No behaviour change. web/src/lib/thesis.ts becomes a re-export; the policy
itself is now worker/src/thesis-policy.ts.

It was in web/src/lib, which was correct while a browser was the only reader.
The orchestrator is about to be a second one — it materialises each child's
followed theses into a file the agent's desk reads, and what it writes has to
be exactly what the public feed publishes: same allowlist, same address
backstop, same fail-closed default.

The worker cannot import from web/src. imports.test.ts forbids @merrymen/*
under worker/src, and web/src is not aliased inward at all. So the choice was
to move the module or to keep a second copy, and a second copy of a
PUBLICATION POLICY is the worst option on the table: the two readers drift into
different ideas of what may be published, and the drift stays invisible until
something private appears on a page.

The module has no imports at all, which is what made this mechanical — it is
pure, DB-free, fetch-free and env-free, exactly as its own header claims.

The alias is registered in BOTH tsconfigs on purpose. web/tsconfig governs the
Next build; the ROOT config governs `tsx --test`, which runs the web tree too.
Registering it in one place only resolves in the app and fails in the tests,
which is how this landed the first time.
