+++ web/tsconfig.json
@@ -22,6 +22,7 @@
       "@merrymen/grant-store": ["../worker/src/grant-store.ts"],
       "@merrymen/command-files": ["../worker/src/command-files.ts"],
       "@merrymen/settings-store": ["../worker/src/settings-store.ts"],
+      "@merrymen/thesis": ["../worker/src/thesis-policy.ts"],
       "@merrymen/llm": ["../worker/src/llm.ts"],
       "@merrymen/settings": ["../worker/src/settings.ts"],
       "@merrymen/home": ["../worker/src/home.ts"],
